% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnbconv.R
\name{dnbconv}
\alias{dnbconv}
\title{Probability Mass Function}
\usage{
dnbconv(
  counts,
  mus,
  ps,
  phis,
  method = c("exact", "moments", "saddlepoint"),
  n.terms = 1000,
  n.cores = 1,
  tolerance = 0.001,
  normalize = TRUE
)
}
\arguments{
\item{counts}{Integer vector of counts over which the convolution is evaluated.}

\item{mus}{Numeric vector of individual mean values}

\item{ps}{Numeric vector of individual probabilities of success.}

\item{phis}{Numeric vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{method}{The method by which to evaluate the PMF. One of "exact", "moments", or "saddlepoint".}

\item{n.terms}{The number of terms to include in the series representation of the exact PMF. Defaults to 1000.}

\item{n.cores}{The number of CPU cores to use in the evaluation. Allows parallelization.}

\item{tolerance}{The acceptable difference between the sum of the K distribution and 1. Defaults to 1E-3.}

\item{normalize}{Boolean. If TRUE, the PMF is normalized to sum to 1.}
}
\value{
A numeric vector of probability densities.
}
\description{
Calculates the probability mass function for the convolution of arbitrary negative binomial random variables.
}
\examples{
dnbconv(counts = 0:500, mus = c(100, 10), phis = c(5, 8), method = "exact")

}
