% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_empirical_sum.R
\name{nb_empirical_sum}
\alias{nb_empirical_sum}
\title{Calculate the empirical sum of arbitrary NB random variables}
\usage{
nb_empirical_sum(mus, phis, ps, n.samp)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities.}

\item{n.samp}{The number of samples of each distribution}
}
\description{
Calculate the empirical sum of arbitrary NB random variables
}
