% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_exact.R
\name{nb_exact}
\alias{nb_exact}
\title{Implements Furman's exact PMF for the evaluation of the sum of arbitrary NB random variables}
\usage{
nb_exact(
  mus,
  phis,
  ps,
  n.terms = 1000,
  counts.start = 0,
  counts.end,
  n.cores = 1
)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities.}

\item{n.terms}{The number of terms to include in the series for evaluating the PMF at a given number of counts. Defaults to 1000.}

\item{counts.start}{The smallest number of counts at which the PMF is to be evaluated. Defaults to 0.}

\item{counts.end}{The largest number of counts at which the PMF is to be evaluated.}

\item{n.cores}{The number of CPU cores to use in the evaluation. Allows parallelization.}
}
\description{
Implements Furman's exact PMF for the evaluation of the sum of arbitrary NB random variables
}
