% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_saddlepoint.R
\name{nb_saddlepoint}
\alias{nb_saddlepoint}
\title{Implements the saddlepoint approximation for the sum of arbitrary NB random variables}
\usage{
nb_saddlepoint(
  mus,
  phis,
  ps,
  counts.start = 0,
  counts.end,
  log.scale = TRUE,
  normalize = TRUE
)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities.}

\item{counts.start}{The smallest number of counts at which the PMF is to be evaluated. Defaults to 0.}

\item{counts.end}{The largest number of counts at which the PMF is to be evaluated.}

\item{log.scale}{Boolean. If TRUE, saddlepoint approximation is done on the log scale.}

\item{normalize}{Boolean. If TRUE, the PMF is re-normalized to sum to 1.}
}
\description{
Adapted from https://stats.stackexchange.com/questions/72479/generic-sum-of-gamma-random-variables/137318#137318
and https://www.martinmodrak.cz/2019/06/20/approximate-densities-for-sums-of-variables-negative-binomials-and-saddlepoint/
}
