% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbconv_params.R
\name{nbconv_params}
\alias{nbconv_params}
\title{Calculates distribution parameters for the convolution of arbitrary negative binomial random variables.}
\usage{
nbconv_params(mus, phis, ps)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities of success.}
}
\value{
A named numeric vector of distribution parameters.
}
\description{
Calculates distribution parameters for the convolution of arbitrary negative binomial random variables.
}
\examples{
nbconv_params(mus = c(100, 10), phis = c(5, 8))

}
