% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnbconv.R
\name{pnbconv}
\alias{pnbconv}
\title{Calculates the CDF for the convolution of arbitrary negative binomial rv's}
\usage{
pnbconv(
  targets,
  mus,
  ps,
  phis,
  method = c("exact", "moments", "saddlepoint"),
  n.terms = 1000,
  n.cores = 1,
  tolerance = 0.001,
  normalize = TRUE
)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{ps}{Vector of individual probabilities.}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{method}{The method by which to evaluate the PMF. One of "exact", "moments", or "saddlepoint".}

\item{n.terms}{The number of terms to include in the series for evaluating the PMF at a given number of counts. Defaults to 1000.}

\item{n.cores}{The number of CPU cores to use in the evaluation. Allows parallelization.}

\item{tolerance}{The acceptable difference between the sum of the K distribution and 1.}

\item{normalize}{Boolean. If TRUE, the PMF is re-normalized to sum to 1.}

\item{counts}{The counts over which the convolution is evaluated. Should be a vector.}
}
\description{
Calculates the CDF for the convolution of arbitrary negative binomial rv's
}
