% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnbconv.R
\name{qnbconv}
\alias{qnbconv}
\title{Quantile Function}
\usage{
qnbconv(
  probs,
  counts,
  mus,
  ps,
  phis,
  method = c("exact", "moments", "saddlepoint"),
  n.terms = 1000,
  n.cores = 1,
  tolerance = 0.001,
  normalize = TRUE
)
}
\arguments{
\item{probs}{Numeric vector of target (cumulative) probabilities.}

\item{counts}{Integer vector of counts over which the PMF is evaluated.}

\item{mus}{Numeric vector of individual mean values}

\item{ps}{Numeric vector of individual probabilities of success.}

\item{phis}{Numeric vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{method}{The method by which to evaluate the PMF. One of "exact", "moments", or "saddlepoint".}

\item{n.terms}{The number of terms to include in the series representation of the exact PMF. Defaults to 1000.}

\item{n.cores}{The number of CPU cores to use in the evaluation. Allows parallelization.}

\item{tolerance}{The acceptable difference between the sum of the K distribution and 1. Defaults to 1E-3.}

\item{normalize}{Boolean. If TRUE, the PMF is normalized to sum to 1.}
}
\value{
A numeric vector of quantiles.
}
\description{
Calculates the quantile function for the convolution of arbitrary negative binomial random variables.
}
\examples{
qnbconv(probs = c(0.05, 0.25, 0.5, 0.75, 0.95), counts = 0:500,
       mus = c(100, 10), phis = c(5, 8), method = "exact")

}
