% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnbconv.R
\name{rnbconv}
\alias{rnbconv}
\title{Generates random samples from the convolution of arbitrary negative binomial rv's}
\usage{
rnbconv(mus, phis, ps, n.samp)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities.}

\item{n.samp}{The number of samples per distribution}
}
\description{
Generates random samples from the convolution of arbitrary negative binomial rv's
}
