% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbconv_params.R
\name{nbconv_params}
\alias{nbconv_params}
\title{Calculates distribution parameters for the convolution of arbitrary negative binomial rv's}
\usage{
nbconv_params(mus, phis, ps)
}
\arguments{
\item{mus}{Vector of individual mean values}

\item{phis}{Vector of individual dispersion parameters. Equivalent to 'size' in dnbinom.}

\item{ps}{Vector of individual probabilities.}
}
\description{
Calculates distribution parameters for the convolution of arbitrary negative binomial rv's
}
